+++ .changeset/web-split-conversation-toc.md
@@ -1,0 +1,5 @@
+---
+"@moonshot-ai/kimi-code": patch
+---
+
+Extract the beta conversation outline (table of contents) into its own component.
